requirements.txt
requests
numpy
flask
keras
Pillow